third_party/WebKit/LayoutTests/dom/domparsing/xmlserializer-double-xmlns-expected.txt

Checks that attributes created setAttributeNS() don't cause two identical xmlns attributes to be serialized

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS doc.documentElement.getAttributeNS("http://www.example.com", "attr") is "value"
PASS parsedDoc.documentElement.getAttributeNS("http://www.example.com", "attr") is "value"
PASS successfullyParsed is true

TEST COMPLETE
